bug caixa de produtos e ordem do caixa de terapias

+++ app/controllers/cashiers_controller.rb
@@ -88,7 +88,7 @@ class CashiersController < ApplicationController
     # Never trust parameters from the scary internet, only allow the white list through.
     def cashier_params
       cp = params.require(:cashier).permit(:client_id, :user_id, :price, :paymment , :paid_value, :paid, :discount, :therapist_value , items_attributes: [:id, :client, :user, :therapy, :description, :_destroy], product_movements_attributes: [:id, :product_id, :kind, :price, :quantity, :user_id, :_destroy], pagamento_caixa_attributes: [:id, :pagamento_id, :value, :_destroy])
-      cp[:therapist_value] = params[:cashier][:therapist_value].to_d
+      cp[:therapist_value] = params[:cashier][:therapist_value].to_f
       cp[:discount] = params[:cashier][:discount].to_i
       cp[:paymment] = params[:cashier][:paymment].to_i
       cp[:paid] = params[:cashier][:paid].to_i
